+++ c#/Katas/GildedRose/src/Refacto/GildedRose.Core/Items/ItemsName.cs
@@ -9,3 +9,4 @@ public enum ItemsName
     ElixirOfMangoust,
     ConjuredManaCake,
 }
+
